% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afttestplot.R
\name{afttestplot_form}
\alias{afttestplot_form}
\title{afttestplot_form}
\usage{
afttestplot_form(object, path, std)
}
\arguments{
\item{object}{is a \code{afttest} fit}

\item{path}{A numeric value specifies the number of approximated processes plotted
The default is set to be 100.}

\item{std}{A character string specifying if the graph is based on the 
unstandardized test statistics or standardized test statistics
The default is set to be "std".}
}
\value{
\code{afttestplot_form} returns a plot based on the \code{form}:
   An object of the functional form test is the form of n by 1 matrix.
   See the documentation of \pkg{ggplot2} and \pkg{gridExtra} for details.
}
\description{
It gives plot for cheking the aft model assumptions.
}
\keyword{internal}
