% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afttestplot.R
\name{afttestplot_omni}
\alias{afttestplot_omni}
\title{afttestplot_omni}
\usage{
afttestplot_omni(object, path, std)
}
\arguments{
\item{object}{is a \code{afttest} fit}

\item{path}{A numeric value specifies the number of approximated processes plotted
The default is set to be 100.}

\item{std}{A character string specifying if the graph is based on the 
unstandardized test statistics or standardized test statistics
The default is set to be "std".}
}
\value{
\code{afttestplot_omni} returns a plot based on the \code{omni}:
   An object of the omnibus test is the form of n by n matrix, 
   some quantiles of x, which are used in weight, are plotted for graphs, 
   i.e. 0\%, 10\%, 25\%, 40\%, 50\%, 60\%, 75\%, 90\%, and 100\% are used.
   See the documentation of \pkg{ggplot2} and \pkg{gridExtra} for details.
}
\description{
It gives plot for cheking the aft model assumptions.
}
\keyword{internal}
