% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_aes}
\alias{export_aes}
\alias{aes}
\title{\code{aes} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{aes}}.
}
\keyword{internal}
