% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_aftsrr}
\alias{export_aftsrr}
\alias{aftsrr}
\title{\code{aftgee} function imported from \code{aftgee}}
\description{
This function is imported from the \code{aftgee} package. See
\code{\link[aftgee]{aftsrr}}.
}
\keyword{internal}
