% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_as.formula}
\alias{export_as.formula}
\alias{as.formula}
\title{\code{as.formula} function imported from \code{stats}}
\description{
This function is imported from the \code{stats} package. See
\code{\link[stats]{as.formula}}.
}
\keyword{internal}
