% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_element_text}
\alias{export_element_text}
\alias{element_text}
\title{\code{element_text} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{element_text}}.
}
\keyword{internal}
