% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_geom_label}
\alias{export_geom_label}
\alias{geom_label}
\title{\code{geom_label} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{geom_label}}.
}
\keyword{internal}
