% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_geom_step}
\alias{export_geom_step}
\alias{geom_step}
\title{\code{geom_step} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{geom_step}}.

This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{geom_step}}.
}
\keyword{internal}
