% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_get_all_vars}
\alias{export_get_all_vars}
\alias{get_all_vars}
\title{\code{get_all_vars} function imported from \code{stats}}
\description{
This function is imported from the \code{stats} package. See
\code{\link[stats]{get_all_vars}}.
}
\keyword{internal}
