% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_ggplot}
\alias{export_ggplot}
\alias{ggplot}
\title{\code{ggplot} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{ggplot}}.
}
\keyword{internal}
