% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_ggtitle}
\alias{export_ggtitle}
\alias{ggtitle}
\title{\code{ggtitle} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{ggtitle}}.
}
\keyword{internal}
