% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_grid.arrange}
\alias{export_grid.arrange}
\alias{grid.arrange}
\title{\code{grid.arrange} function imported from \code{gridExtra}}
\description{
This function is imported from the \code{gridExtra} package. See
\code{\link[gridExtra]{grid.arrange}}.
}
\keyword{internal}
