% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_optim}
\alias{export_optim}
\alias{optim}
\title{\code{optim} function imported from \code{stats}}
\description{
This function is imported from the \code{stats} package. See
\code{\link[stats]{optim}}.
}
\keyword{internal}
