% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_quantile}
\alias{export_quantile}
\alias{quantile}
\title{\code{quantile} function imported from \code{stats}}
\description{
This function is imported from the \code{stats} package. See
\code{\link[stats]{quantile}}.
}
\keyword{internal}
