% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_scale_y_continuous}
\alias{export_scale_y_continuous}
\alias{scale_y_continuous}
\title{\code{scale_y_continuous} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{scale_y_continuous}}.
}
\keyword{internal}
