% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_theme}
\alias{export_theme}
\alias{theme}
\title{\code{theme} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{theme}}.
}
\keyword{internal}
