% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_theme_minimal}
\alias{export_theme_minimal}
\alias{theme_minimal}
\title{\code{theme_minimal} function imported from \code{ggplot2}}
\description{
This function is imported from the \code{ggplot2} package. See
\code{\link[ggplot2]{theme_minimal}}.
}
\keyword{internal}
