% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_r.R
\name{plotting_link}
\alias{plotting_link}
\title{plotting_link}
\usage{
plotting_link(result, path, std)
}
\arguments{
\item{result}{For function afttestplot, the only required argument is
afttestresult based on the result from the afttest function. Whatever
the testtype of the result is, it automatically gives the corresponding
graph.}

\item{path}{The argument path is for the number of the simulation paths
that is plotted in the graph. Therefore it needs to be equal or less than
the number of paths used in by afttest function, otherwise it is given as
the number of paths used in by afttest function. The default is set to be 100.}

\item{std}{The option for the argument std is "unstd" and "std". In this 
argument, "std" is the default.}
}
\value{
Basically, a graph from the afttestplot is based on the packages
ggplot2 (Wickham, 2009) and gridExtra (Auguie, 2017). It offers a graph that
y-axis is the test statistics and x-axis represents the rank of the subjects
ordered by time transformed residual.
}
\description{
It gives plot for cheking the aft model assumptions.
}
