% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afttest.R
\name{afttest.formula}
\alias{afttest.formula}
\title{Model Diagnostics for AFT Models using Formulas}
\usage{
\method{afttest}{formula}(
  object,
  data,
  npath = 200,
  testType = "omnibus",
  estMethod = "rr",
  eqType = "ns",
  covTested = 1,
  npathsave = 50,
  linApprox = TRUE,
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{A formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.
See the documentation of \code{lm}, \code{coxph} and \code{formula} for details.}

\item{data}{An optional data frame in which to interpret the variables occurring 
in the formula.}

\item{npath}{An integer value specifying the number of approximated processes.
The default is given by 200.}

\item{testType}{A character string specifying the type of the test.
  The following are permitted:
  \describe{
    \item{\code{omnibus}}{an omnibus test}
    \item{\code{link}}{a link function test}
    \item{\code{covForm}}{a functional form of a covariate}
}}

\item{estMethod}{A character string specifying the type of the estimator used.
  The readers are referred to the \pkg{aftgee} package for details.
  The following are permitted:
  \describe{
    \item{\code{ls}}{Least-Squares Approach for Accelerated Failure Time 
    with Generalized Estimating Equation}
    \item{\code{rr}}{Accelerated Failure Time with Smooth Rank Regression}
}}

\item{eqType}{A character string specifying the type of the 
  estimating equation used to obtain the regression parameters.
  The readers are referred to the \pkg{aftgee} package for details.
  The following are permitted:
  \describe{
    \item{\code{ns}}{Regression parameters are estimated by directly solving 
    the nonsmooth estimating equations.}
    \item{\code{is}}{Regression parameters are estimated by directly solving 
    the induced-smoothing estimating equations.}
}}

\item{covTested}{A character string specifying the covariate which will be tested.
The argument \code{covTested} is necessary only if \code{testType} is 
\code{covForm}. The default option for \code{covTested} is given by "1", which 
represents the first covariate in the formula argument.}

\item{npathsave}{An integer value specifying the number of paths saved among all the paths.
The default is given by 50. Note that it requires a lot of memory if saving all
sampled paths (N by N matrix for each npath, and so npath*N*N elements).}

\item{linApprox}{A logical value. If \code{TRUE}, the multiplier bootstrap is 
computed using the asymptotic linear approximation, which is significantly 
faster. If \code{FALSE}, the estimating equations are solved numerically for 
each bootstrap replication. Defaults to \code{TRUE}.}

\item{seed}{An optional integer specifying the random seed for reproducibility.}

\item{...}{Other arguments passed to methods.}
}
\description{
Model Diagnostics for AFT Models using Formulas
}
